Visit the rightmost child first, for performance reasons.
Visit the rightmost child first, to save execution time in case we don't have to loop over left children, which is in 50% of cases on average.

Perhaps the order isn't intuitive, to first go right and then left, but it saves time.

We couldn't reverse the order in only one function, for some reason,`select_columns_in_order_rec()`.